requirements.txt
Flask==3.1.0
rembg==2.0.62
Pillow==11.1.0
onnxruntime==1.20.1
